+++ pkg/grid/mountain.go
@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-func TerrainGenerator(x, y int, xFrequency, yFrequency, gain float64, minHeight uint16, forceBaseLand bool) [][]Element {
+func TerrainGenerator(x, y int, xFrequency, yFrequency, gain float64, minHeight int, forceBaseLand bool) [][]Element {
 	groundHeight := [][]Element{}
 
 	for i := 0; i < x; i++ {
@@ -37,7 +37,7 @@ func TerrainGenerator(x, y int, xFrequency, yFrequency, gain float64, minHeight 
 			xHeight := (gain * math.Sin(xNormalizedValue*math.Pi)) + gain
 			yHeight := (gain * math.Sin(yNormalizedValue*math.Pi)) + gain
 
-			heightAvg := uint16((xHeight + yHeight) / 2.0)
+			heightAvg := int((xHeight + yHeight) / 2.0)
 
 			groundHeight[i][j].Height = heightAvg + minHeight
 
@@ -51,7 +51,7 @@ func TerrainGenerator(x, y int, xFrequency, yFrequency, gain float64, minHeight 
 	return groundHeight
 }
 
-func MountainGenerator(x, y int, gain float64, minHeight uint16) [][]Element {
+func MountainGenerator(x, y int, gain float64, minHeight int) [][]Element {
 	xFrequency := 2.0
 	yFrequency := 2.0
 
@@ -79,10 +79,10 @@ func MountainGenerator(x, y int, gain float64, minHeight uint16) [][]Element {
 			xHeight := (gain * math.Sin(xNormalizedValue*math.Pi)) + gain
 			yHeight := (gain * math.Sin(yNormalizedValue*math.Pi)) + gain
 
-			heightAvg := uint16((xHeight + yHeight) / 2.0)
+			heightAvg := int((xHeight + yHeight) / 2.0)
 
-			if heightAvg > uint16(gain) {
-				groundHeight[i][j].Height = heightAvg - uint16(gain) + minHeight
+			if heightAvg > int(gain) {
+				groundHeight[i][j].Height = heightAvg - int(gain) + minHeight
 			} else {
 				groundHeight[i][j].Height = 0
 			}
